Distribute more tests
